Integrate Session API (#347)
OKTA-453788 Integrate Session API

+++ test/delete-resources.js
@@ -32,42 +32,82 @@ async function cleanAuthorizationServers() {
 
 async function cleanApplications() {
   await (await client.listApplications()).each(async application => {
-    try {
-      [
-        'Okta Admin Console',
-        'Okta Dashboard',
-        'Okta Browser Plugin',
-        'Node SDK Service App',
-        'Bacon Service Client'
-      ].includes(application.label) ?
-        console.log(`Skipped application to remove ${application.label}`) :
+    const canDelete = ![
+      'Okta Admin Console',
+      'Okta Dashboard',
+      'Okta Browser Plugin',
+      'Node SDK Service App',
+      'Bacon Service Client'
+    ].includes(application.label) && application.label.startsWith('node-sdk: ');
+    if (canDelete) {
+      try {
         await utils.removeAppByLabel(client, application.label);
-    } catch (err) {
-      console.error(err);
+      } catch (err) {
+        console.error(err);
+      }
+    } else {
+      console.log(`Skipped application to remove ${application.label}`);
     }
   });
 }
 
 async function cleanTestUsers() {
   await (await client.listUsers()).each(async user => {
-    try {
-      (user.profile.email.endsWith('okta.com')) ?
-        console.log(`Skipped user to remove ${user.profile.email}`) :
+    const canDelete = !user.profile.email.endsWith('okta.com');
+    if (canDelete) {
+      try {
         await utils.deleteUser(user, client);
-    } catch (err) {
-      console.error(err);
+      } catch (err) {
+        console.error(err);
+      }
+    } else {
+      console.log(`Skipped user to remove ${user.profile.email}`);
     }
   });
 }
 
 async function cleanTestGroups() {
   await (await client.listGroups()).each(async (group) => {
-    try {
-      (group.profile.name === 'Everyone') ?
-        console.log(`Skipped group to remove ${group.profile.name}`) :
+    const canDelete = group.profile.name !== 'Everyone' && group.profile.name.startsWith('node-sdk:');
+    if (canDelete) {
+      try {
         await utils.cleanupGroup(client, group);
-    } catch (err) {
-      console.error(err);
+      } catch (err) {
+        console.error(err);
+      }
+    } else {
+      console.log(`Skipped group to remove ${group.profile.name}`);
+    }
+  });
+}
+
+async function cleanTestPolicies() {
+  await (await client.listPolicies({ type: 'OKTA_SIGN_ON' })).each(async policy => {
+    const canDelete = policy.name.startsWith('node-sdk:');
+    if (canDelete) {
+      try {
+        await client.deletePolicy(policy.id);
+      } catch (err) {
+        console.error(err);
+      }
+    } else {
+      console.log(`Skipped policy to remove ${policy.name}`);
+    }
+  });
+}
+
+async function cleanTestIdps() {
+  await (await client.listIdentityProviders()).each(async idp => {
+    const canDelete = idp.name.startsWith('node-sdk:');
+    if (canDelete) {
+      try {
+        await client.deactivateIdentityProvider(idp.id);
+        await client.deleteIdentityProvider(idp.id);
+      } catch (err) {
+        console.error(err);
+      }
+    } else {
+      console.log(`Skipped IDP to remove ${idp.name}`);
     }
   });
 }
@@ -79,5 +119,7 @@ describe('Clean', () => {
     await cleanTestGroups();
     await cleanApplications();
     await cleanInlineHooks();
+    await cleanTestPolicies();
+    await cleanTestIdps();
   });
 });
